mqtt-publisher-json/app/requirements.txt
paho-mqtt==2.1.0
ujson==5.10.0